Fixed error Profiler.

+++ examples/Renderer/CUDAWindow.cpp
@@ -127,13 +127,13 @@ void TriangleWindow::ProfilerBegin() {
 	QueryPerformanceCounter(&m_t1);
 }
 
-void TriangleWindow::ProfilerEnd(int numRays) {
+void TriangleWindow::ProfilerEnd(long long int numRays) {
 	double dff = m_tc.QuadPart;
 	QueryPerformanceCounter(&m_t2);
 	double s = (m_t2.QuadPart - m_t1.QuadPart) / dff;
 	printf("render time is %.3fs\n", s);
-	int rayPerSecond = numRays / s;
-	printf("path per second %d rays/s\n", rayPerSecond);
+	long long int rayPerSecond = numRays / s;
+	printf("path per second %lld rays/s\n", rayPerSecond);
 }
 
 void TriangleWindow::createVBO(GLuint* vbo)
@@ -411,7 +411,7 @@ void TriangleWindow::paintGL()
 	if(framenumber % m_interval == 0){
 		const int depth = 4;
 		const int samp = 1;
-		int numRays = width() * height() * depth * samp * framenumber;
+		long long int numRays = (long long)width() * (long long)height() * (long long)(depth * samp * framenumber);
 		ProfilerEnd(numRays);
 	}
